Increased specificity on p selector to limit reach
With how the paragraph tag was being selected and modified in the TeamCard CSS file, we were making all paragraphs on the entire app take on those styles.
To prevent this, I just made sure to only target paragraphs within the .player_card class.
In the future, maybe we should use CSS Modules to prevent these types of unintended behavior.