Add fetching of mock Match data from our backend!
Rather than use a hard-coded object for Match data, the MatchSummary now grabs its Match data from our backend server based on the ID parameter in the URL (/matches/:id).
The server is also just grabbing the data from a locally-stored JSON file. But this is one step closer to structuring this component in a more flexible way.