Began implementing the profile page
I have begun implementing the profile page. So far I have split it into 3 components, PlayerStats, PlayerGods, and ClassStats. Each has their own subcomponents.
Styling is temporary, need to figure out what format the data will be coming in from the api.

Sidenote: the stat class is just an array of [name, value] so that I can make it easier to construct 2d arrays that containt the name and value of several stats